convert -88 and empty string to NA
I want to make sure the phabmetrics package is not doing any calculations with -88s. I also want to make sure it recognizes empty strings as NA in the VariableResult column. I changed some of the checker code to deal with how it deals with NA and Null values, and I want to make sure it doesn't throw off metrics